Refactor QRParser: move to separate file, add XML documentation in Russian, replace marker logic with GS1 parser, add .gitignore for IDE files
