+++ src/tabalon.ismp.crpt/csp/QRParserTest/QrDataMatrixTests.cs
@@ -1,4 +1,5 @@
-﻿using System;
+using System;
+using TbkIsmpCrptApi;
 using TbkIsmpCrptApi.Controllers;
 using Xunit;
